+++ lib/pages/card_page.dart
@@ -40,8 +40,9 @@ class CardsState extends State<Cards> {
                         bottomLeft: Radius.circular(22),
                         bottomRight: Radius.circular(22))),
               ),
-              Padding(
-                padding: const EdgeInsets.only(top: 70.0, left: 32),
+              Positioned(
+                top: MediaQuery.of(context).size.width * 0.2,
+                left: MediaQuery.of(context).size.width * 0.1,
                 child: Container(
                   height: MediaQuery.of(context).size.height * 0.6,
                   width: MediaQuery.of(context).size.width * 0.5,
@@ -51,18 +52,12 @@ class CardsState extends State<Cards> {
                 ),
               ),
               Positioned(
-                top: MediaQuery.of(context).size.width * 0.6,
+                top: MediaQuery.of(context).size.width * 0.55,
                 child: Container(
-                  height: MediaQuery.of(context).size.height * 0.1,
+                  height: MediaQuery.of(context).size.height * 0.2,
                   width: MediaQuery.of(context).size.width * 0.8,
                   child: Transform(
-                    child: BarcodeWidget(
-                      style: TextStyle(fontSize: 20),
-                      height: 100,
-                      width: 350,
-                      barcode: Barcode.ean13(),
-                      data: args.cards["barecode"],
-                    ),
+                    child: getBarcodeWidget(args.cards),
                     alignment: FractionalOffset.center,
                     transform: new Matrix4.identity()
                       ..rotateZ(90 * 3.1415927 / 180),
@@ -70,11 +65,11 @@ class CardsState extends State<Cards> {
                 ),
               ),
               Positioned(
-                top: MediaQuery.of(context).size.width * 0.7,
-                right: MediaQuery.of(context).size.width * 0.3,
+                right: MediaQuery.of(context).size.width * 0.23,
                 child: Transform(
                   child: Container(
-                    width: MediaQuery.of(context).size.width * 0.4,
+                    width: MediaQuery.of(context).size.width * 0.6,
+                    height: MediaQuery.of(context).size.height * 0.85,
                     alignment: Alignment.center,
                     child: Text(
                       args.cards["label"],
@@ -86,8 +81,9 @@ class CardsState extends State<Cards> {
                     ..rotateZ(91 * 3.1415927 / 180),
                 ),
               ),
-              Padding(
-                padding: const EdgeInsets.only(left: 240, top: 10),
+              Positioned(
+                top: MediaQuery.of(context).size.width * 0.05,
+                left: MediaQuery.of(context).size.width * 0.6,
                 child: Container(
                   height: 25,
                   width: 25,
@@ -101,6 +97,158 @@ class CardsState extends State<Cards> {
         ],
       ),
     );
+  }
+
+  BarcodeWidget getBarcodeWidget(Map<String, dynamic> _card) {
+    switch (_card["format"]) {
+      case "ean13":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.ean13(),
+          data: _card["barecode"],
+        );
+        break;
+      case "code128":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.code128(),
+          data: _card["barecode"],
+        );
+        break;
+      case "code39":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.code39(),
+          data: _card["barecode"],
+        );
+        break;
+      case "code93":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.code93(),
+          data: _card["barecode"],
+        );
+        break;
+      case "dataMatrix":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.dataMatrix(),
+          data: _card["barecode"],
+        );
+        break;
+      case "ean2":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.ean2(),
+          data: _card["barecode"],
+        );
+        break;
+      case "ean5":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.ean5(),
+          data: _card["barecode"],
+        );
+        break;
+      case "ean8":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.ean8(),
+          data: _card["barecode"],
+        );
+        break;
+      case "gs128":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.gs128(),
+          data: _card["barecode"],
+        );
+        break;
+      case "isbn":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.isbn(),
+          data: _card["barecode"],
+        );
+        break;
+      case "itf":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.itf(),
+          data: _card["barecode"],
+        );
+        break;
+      case "itf14":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.itf14(),
+          data: _card["barecode"],
+        );
+        break;
+      case "pdf417":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.pdf417(),
+          data: _card["barecode"],
+        );
+        break;
+      case "qr":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.qrCode(),
+          data: _card["barecode"],
+        );
+        break;
+      case "rm4scc":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.rm4scc(),
+          data: _card["barecode"],
+        );
+        break;
+      case "aztec":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.aztec(),
+          data: _card["barecode"],
+        );
+        break;
+      case "upce":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.upcE(),
+          data: _card["barecode"],
+        );
+        break;
+      case "upca":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.upcA(),
+          data: _card["barecode"],
+        );
+        break;
+      case "codabar":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.codabar(),
+          data: _card["barecode"],
+        );
+        break;
+      case "telepen":
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.telepen(),
+          data: _card["barecode"],
+        );
+        break;
+      default:
+        return BarcodeWidget(
+          style: TextStyle(fontSize: 20),
+          barcode: Barcode.ean13(),
+          data: _card["barecode"],
+        );
+        break;
+    }
   }
 
   goToLoyaltyCardsPage() {