+++ CHANGELOG.md
@@ -51,6 +51,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Widget_share_contact : contact name overflow
 - Create_list : wishlist creation screen overflow
 - Authentication-page : Account creation page not scrollable
+- Loyalty Cards Page : Changed floating button icon
+- Widget_item : Product designation overflow
+- Share_one and Mainpage : Wishlist share process indicator visible on short process
+- Card_page : Big loyalty card render problem
+- Loyalty Cards Page : User can now scan different barcode types
 
 ### Security